+++ parent_algorithm.cpp
@@ -14,16 +14,16 @@ Parent_Algorithm::Parent_Algorithm()
 
 }
 
-int16_t*** Parent_Algorithm::Generate_Initial_Cube() {
+int32_t*** Parent_Algorithm::Generate_Initial_Cube() {
 
     //Создаём динамический массив. Вместо (30) подставить numCubes
-    voxels = new int16_t** [numCubes];
+    voxels = new int32_t** [numCubes];
     assert(voxels);
     for (int i = 0; i < numCubes; i++) {
-        voxels[i] = new int16_t* [numCubes];
+        voxels[i] = new int32_t* [numCubes];
         assert(voxels[i]);
         for (int j = 0; j < numCubes; j++) {
-            voxels[i][j] = new int16_t[numCubes];
+            voxels[i][j] = new int32_t[numCubes];
             assert(voxels[i][j]);
         }
     }
